add readme and documentation
Add information to `lib.rs`, a readme, and automation to ensure they stay in
sync via `cargo sync-rdme` (I did a survey of them a month or so ago, and this
is my favorite one).

Some references within the readmes are currently a little broken -- they should
be fixed once the crates are on crates.io.

Reviewers: andrewjstone

Reviewed By: andrewjstone

Pull Request: https://github.com/oxidecomputer/daft/pull/30