[3/n] add a function to strip a reference from a leaf node
While working on some other code, I realized that having this is quite
convenient. It won't ordinarily be created via `Diffable::diff`, but it can be
via a leaf annotation or via an upcoming change to how maps work.

I struggled to name this for a while:

* `deref` isn't right because this doesn't return a reference
* `as_ref` isn't right for the same reason
* `copied` does describe this, but this only works for references, not
  arbitrary `T: Copy`.

So I settled on `unref`, which is uncommon but hopefully clear enough.

Reviewers: andrewjstone

Reviewed By: andrewjstone

Pull Request: https://github.com/oxidecomputer/daft/pull/33