Impl Diffable for NonZero with 'leaf!' (#68)
I noticed that `std::num::NonZero<T>` didn't impl `Diffable`, which I
would like to have for a project I'm working on. This simply uses the
`leaf!{}` macro to implement `Diffable` for each of the `NonZero*`
types.

I would have liked to do `impl<T: Diffable> Diffable for NonZero<T>` but
that requires us to add the `T: NonZeroablePrimitive` bound to the impl
(since that is a bound on the `NonZero` type itself), but we can't do
that since that trait isn't stable yet.